Refactor: Decoupled team members into seperate 'Team' component
